consolidated the abstract classes into two files: mixins and base
