+++ apps/site/models/__init__.py
@@ -1,12 +1,11 @@
 # abstract
-from localground.apps.site.models.abstract.base import Base
-from localground.apps.site.models.abstract.audit import BaseAudit
-from localground.apps.site.models.abstract.named import BaseNamed
-from localground.apps.site.models.abstract.media import BaseMediaMixin, BaseUploadedMedia
-from localground.apps.site.models.abstract.geometry import ExtentsMixin, PointMixin
-from localground.apps.site.models.abstract.mixins import ExtrasMixin, ProjectMixin, BaseGenericRelationMixin
+from localground.apps.site.models.abstract.base import Base, BaseAudit, \
+    BaseNamed, BaseNamedMedia, BaseUploadedMedia
+from localground.apps.site.models.abstract.mixins import ExtentsMixin, \
+    PointMixin, ExtrasMixin, ProjectMixin, BaseGenericRelationMixin, \
+    BaseMediaMixin
 
-#layers
+# layers
 from localground.apps.site.models.layer import Layer
 
 # lookups
